pyproject.toml
[project]
name = "pipeline-project"
version = "0.1.0"
description = "A modular asynchronous pipeline for processing encrypted tarballs"
authors = [
    {name = "Cliff Wulfman", email = "cwulfman@princeton.edu"},
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rich",
    "pyyaml>=6.0.2",
    "google-api-python-client>=2.174.0",
    "oauth2client>=4.1.3",
    "httpx>=0.28.1",
    "boto3>=1.39.11",
    "google-auth-oauthlib>=1.2.2",
    "dotenv>=0.9.9",
]

[tool.pdm]
package-dir = "src"

[tool.pdm.scripts]
orchestrator = "python src/pipeline/orchestrator.py"
viewer = "python src/pipeline/token_log_viewer.py"
# lint = "basedpyright"
lint="ruff"
test = "pytest"

[lint]
select = ["E", "F", "I"]
ignore = ["E501"]
line-length = 88

[tool.ruff]




[tool.basedpyright]
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.10"
useLibraryCodeForTypes = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
dev = [
    "basedpyright>=1.28.5",
    "pytest-datadir>=1.8.0",
    "mypy>=1.17.0",
    "isort>=6.0.1",
    "ruff>=0.13.1",
]
